pyproject.toml
[tool.poetry]
name = "langchain-basic"
version = "0.1.0"
description = ""
authors = ["Suraj"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
black = "^24.10.0"
pandas = "^2.2.3"
matplotlib = "^3.10.0"
seaborn = "^0.13.2"
scikit-learn = "^1.6.1"
jupyter = "^1.1.1"
langchain = "^0.3.14"
langchain-ollama = "^0.2.2"
langchain-community = "^0.3.14"
pypdf = "^5.1.0"
chromadb = "^0.6.2"
langchain-openai = "^0.3.5"
wikipedia = "^1.4.0"
googlesearch-python = "^1.3.0"
langchain-google-genai = "^2.0.9"
google-search-results = "^2.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
